before handling user crud

+++ erp-backend/Modules/Core/app/Repositories/Branch/BranchRepository.php
@@ -1,35 +1,44 @@
-﻿<?php
+<?php
 
 namespace Modules\Core\Repositories\Branch;
 
-use App\Http\Responses\ApiResponse;
 use App\Traits\API;
-use App\Repositories\BaseRepository;
 use Illuminate\Support\Facades\File;
 use Spatie\MediaLibrary\MediaCollections\Models\Media;
 use Modules\Core\Models\Branch\Branch;
 use Modules\Core\Repositories\Branch\BranchInterface;
 use Modules\Core\Resources\Branch\BranchResource;
 
-class BranchRepository extends BaseRepository implements BranchInterface
+class BranchRepository implements BranchInterface
 {
-    public function getModel()
+    public function getModel(): \Illuminate\Database\Eloquent\Model
     {
         return new Branch();
     }
 
 
 
-    public function index($request): \Illuminate\Http\JsonResponse
+    public function index($request, $filter): \Illuminate\Http\JsonResponse
     {
-        $perPage = $request['per_page'] ?? config('myConfig.paginationCount');
-        $data    = $perPage == -1 ? $this->getModel()->search($request['search'])->orderBy('created_at', 'desc')->get() : $this->getModel()->search($request['search'])->orderBy('created_at', 'desc')->paginate($perPage);
+        $perPage    = $request['per_page'] ?? config('myConfig.paginationCount');
+        $collection = $this->getModel()->ordering($request->ordering)->filter($filter);
+        $data       = $perPage == -1 ? $collection->where('status', 1)->get() : $collection->paginate($perPage);
 
         return (new API)
-            ->isOk(__('Countries'))
-            ->setData($perPage == -1 ? BranchResource::collection($data) : (new API)->api_model_set_paginate(BranchResource::collection($data) ,$data))
+            ->isOk(__('Branches'))
+            ->setData($perPage == -1 ? BranchResource::collection($data) : (new API)->api_model_set_paginate(BranchResource::collection($data), $data))
             ->build();
+    }
 
+
+
+    public function show($id, array $with = [])
+    {
+        $branch = $this->getModel()->with($with)->findOrFail($id);
+        return (new API)
+            ->isOk(__('Branch Data'))
+            ->setData(BranchResource::make($branch))
+            ->build();
     }
 
 
@@ -39,17 +48,17 @@ class BranchRepository extends BaseRepository implements BranchInterface
         try {
             $branch = $this->getModel()->create($request->validated());
 
-            //save image with branch object
+            //store single image
             if ($request->hasFile('image')) {
-                $branch->addMediaFromRequest('image')->toMediaCollection('branch');
+                $branch->clearMediaCollection('branch'); //delete old record from database
+                $image = $request->file('image');
+                $this->uploadMedia($branch, 'branch', $image);
             }
-
+            //store multi images
             if ($request->hasFile('images')) {
-                if ($images = $request->file('images')) {
-                    $branch->clearMediaCollection('images');
-                    foreach ($images as $image) {
-                        $branch->addMedia($image)->toMediaCollection('images');
-                    }
+                $branch->clearMediaCollection('branch_images'); //delete old record from database
+                foreach($request->file('images') as $image) {
+                    $this->uploadMedia($branch, 'branch_images', $image);
                 }
             }
 
@@ -66,32 +75,25 @@ class BranchRepository extends BaseRepository implements BranchInterface
 
 
 
-    public function show($branch)
-    {
-        return (new API)
-            ->isOk(__('Branch Data'))
-            ->setData(BranchResource::make($branch))
-            ->build();
-    }
-
-
-
-    public function update($branch ,$request)
+    public function update($id, $request)
     {
         try {
+            $branch = $this->getModel()->findOrFail($id);
             $branch->update($request->validated());
 
-            //save new image with branch object and delete old image
+            //update single image
             if ($request->hasFile('image')) {
-                $file_name = $branch->getMedia('branch')->last()->file_name;
-                $img_id = $branch->getMedia('branch')->last()->id;
-                if($img_id && $file_name) {
-                    if (File::exists(public_path('storage/'. $img_id .'/'.$file_name))) {
-                        unlink(public_path('storage/' . $img_id .'/'.$file_name));
-                    }
+                $branch->clearMediaCollection('branch'); //delete old record from database
+                $image = $request->file('image');
+                $this->uploadMedia($branch, 'branch', $image);
+            }
+            //update multi images
+            if ($request->hasFile('images')) {
+                $branch->clearMediaCollection('branch_images'); //delete old record from database
+                foreach($request->file('images') as $image)
+                {
+                    $this->uploadMedia($branch, 'branch_images', $image);
                 }
-                Media::find($branch->getMedia('branch')->last()->id)->delete();
-                $branch->addMediaFromRequest('image')->toMediaCollection('branch');
             }
 
             return (new API)
@@ -107,22 +109,63 @@ class BranchRepository extends BaseRepository implements BranchInterface
 
 
 
-    public function destroy($branch)
+    public function destroy($id)
     {
-        $file_name = $branch->getMedia('branch')->last()->file_name ?? null;
-        $img_id    = $branch->getMedia('branch')->last()->id ?? null;
-        if($img_id && $file_name) {
-            if (File::exists(public_path('storage/'. $img_id .'/'.$file_name))) {
-                unlink(public_path('storage/' . $img_id .'/'.$file_name));
+        try {
+            $branch = $this->getModel()->findOrFail($id);
+
+            //delete old media if exist
+            $singleMedia = $branch->getMedia('branch')->first();
+            $multiMedia  = $branch->getMedia('branch_images')->all();
+            if($singleMedia) {
+                $branch->clearMediaCollection('branch'); //delete old record from database
+                $file_name = $singleMedia->file_name;
+                $img_id    = $singleMedia->id;
+                if($img_id && $file_name) {
+                    //remove files from project
+                    if (File::exists(public_path('storage/'. $img_id .'/'.$file_name)))
+                    {
+                        unlink(public_path('storage/' . $img_id .'/'.$file_name));
+                    }
+                }
             }
+            if($multiMedia) {
+                $branch->clearMediaCollection('branch_images'); //delete old record from database
+                foreach($multiMedia as $media) {
+                    $file_name = $media->file_name;
+                    $img_id    = $media->id;
+                    if($img_id && $file_name) {
+                        //remove files from branch
+                        if (File::exists(public_path('storage/'. $img_id .'/'.$file_name))) {
+                            unlink(public_path('storage/' . $img_id .'/'.$file_name));
+                        }
+                    }
+                }
+            }
+
+            $branch->delete();
+
+            return (new API)
+                ->isOk(__('Destroyed Successfully'))
+                ->build();
         }
-        if($img_id) {
-            Media::find($branch->getMedia('branch')->last()->id)->delete();
+        catch (\Exception $e) {
+            return (new API)
+                ->isError('An Error occured')
+                ->setStatus(500)
+                ->build();
         }
-        $branch->delete();
+    }
+
+
+
+    public function changeStatus($id, $request)
+    {
+        $branch = $this->getModel()->findOrFail($id);
+        $branch->update(['status' => $request->status]);
 
         return (new API)
-            ->isOk(__('Destroyed Successfully'))
+            ->isOk(__('Status Changed Successfully'))
             ->build();
     }
 }
